+++ src/components/app/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ColorRandomizer from '../../containers/Color-Randomizer';
+import ColorRandomizer from '../../containers/ColorRandomizer.jsx';
 
 export default function App() {
   return <ColorRandomizer />;
