+++ src/components/colorDisplay/colorDisplay.css
