+++ src/components/colorDisplay/colorDisplay.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './color-display.css';
+import './colorDisplay.css';
 
 const Display = ({ color }) => {
   return <div style={{ backgroundColor: color }}></div>;
